# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Espaço Sinapse — API administrativa",
    "description": "Sessão HttpOnly. Obtenha GET /api/v1/auth/csrf e envie seu token no headerName retornado em operações de escrita. Após login, obtenha novo CSRF. ADMIN administra cadastros; RECEPCAO gerencia pacientes e agenda. Datas usam offset ISO-8601. Atualizações exigem version. Erros usam ProblemDetail; conflitos retornam 409, incluindo conflictingAppointmentIds quando aplicável. API pública expõe somente campos aprovados.",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8081",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "sessionCookie": []
    }
  ],
  "paths": {
...
    "/api/v1/patients/{id}": {
      "get": {
        "tags": [
          "patient-controller"
        ],
        "operationId": "get_3",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PatientDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "patient-controller"
        ],
        "operationId": "update_3",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "description": "Token retornado por GET /api/v1/auth/csrf. Conserve o cookie da mesma sessão; obtenha um novo token após login.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PatientDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "patient-controller"
        ],
        "operationId": "delete_3",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "description": "Token retornado por GET /api/v1/auth/csrf. Conserve o cookie da mesma sessão; obtenha um novo token após login.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/v1/clinic-settings": {
      "get": {
        "tags": [
          "clinic-controller"
        ],
        "operationId": "settings",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "clinic-controller"
        ],
        "operationId": "update_4",
        "parameters": [
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "description": "Token retornado por GET /api/v1/auth/csrf. Conserve o cookie da mesma sessão; obtenha um novo token após login.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/patients": {
      "get": {
        "tags": [
          "patient-controller"
        ],
        "operationId": "list_3",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": ""
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageDtoPatientDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "patient-controller"
        ],
        "operationId": "create_3",
        "parameters": [
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "description": "Token retornado por GET /api/v1/auth/csrf. Conserve o cookie da mesma sessão; obtenha um novo token após login.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PatientDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/patients/{id}/active": {
      "patch": {
        "tags": [
          "patient-controller"
        ],
        "operationId": "active_3",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "description": "Token retornado por GET /api/v1/auth/csrf. Conserve o cookie da mesma sessão; obtenha um novo token após login.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActiveInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PatientDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/audit-events": {
      "get": {
        "tags": [
          "clinic-controller"
        ],
        "operationId": "audit",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageDtoAuditDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PatientInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 160,
            "minLength": 0
          },
          "birthDate": {
            "type": "string",
            "format": "date"
          },
          "phone": {
            "type": "string",
            "maxLength": 32,
            "minLength": 0
          },
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 254,
            "minLength": 0
          },
          "guardianName": {
            "type": "string",
            "maxLength": 160,
            "minLength": 0
          },
          "guardianRelationship": {
            "type": "string",
            "maxLength": 80,
            "minLength": 0
          },
          "guardianPhone": {
            "type": "string",
            "maxLength": 32,
            "minLength": 0
          },
          "version": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "birthDate",
          "name",
          "phone"
        ]
      },
      "PatientDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "birthDate": {
            "type": "string",
            "format": "date"
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "guardianName": {
            "type": "string"
          },
          "guardianRelationship": {
            "type": "string"
          },
          "guardianPhone": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          },
          "version": {
            "type": "integer",
            "format": "int64"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "SettingsInput": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "maxLength": 160,
            "minLength": 0
          },
          "description": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 0
          },
          "phone": {
            "type": "string",
            "maxLength": 32,
            "minLength": 0
          },
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 254,
            "minLength": 0
          },
          "whatsapp": {
            "type": "string",
            "maxLength": 32,
            "minLength": 0
          },
          "publicAddress": {
            "type": "string",
            "maxLength": 500,
            "minLength": 0
          },
          "addressConfirmed": {
            "type": "boolean"
          },
          "version": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "description",
          "displayName",
          "email",
          "phone",
          "version",
          "whatsapp"
        ]
      },
      "SettingsDto": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "whatsapp": {
            "type": "string"
          },
          "publicAddress": {
            "type": "string"
          },
          "addressConfirmed": {
            "type": "boolean"
          },
          "version": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "ActiveInput": {
        "type": "object",
        "properties": {
          "active": {
            "type": "boolean"
          },
          "version": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "active",
          "version"
        ]
      },
...
      "PageDtoPatientDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PatientDto"
            }
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "AuditDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "actorId": {
            "type": "string",
            "format": "uuid"
          },
          "action": {
            "type": "string"
          },
          "entityType": {
            "type": "string"
          },
          "entityId": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "PageDtoAuditDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuditDto"
            }
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
    },
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "name": "JSESSIONID",
        "in": "cookie"
      }
    }
  }
}